+++ src/RcsbGroupWeb/RcsbGroupView/RcsbGroupDisplay.tsx
@@ -12,6 +12,7 @@ import {RcsbStatsChartComponent} from "./RcsbGroupChart/RcsbStatsChartComponent"
 import {createRoot} from "react-dom/client";
 import {ResidueChartInterface} from "./RcsbResidueChart/ResidueChartTools/ResidueChartTools";
 import {RcsbResidueChartComponent} from "./RcsbResidueChart/RcsbResidueChartComponent";
+import {ChartDisplayConfigInterface} from "../../RcsbChartWeb/RcsbChartComponent/ChartConfigInterface";
 
 
 export class RcsbGroupDisplay {
@@ -42,9 +43,9 @@ export class RcsbGroupDisplay {
         );
     }
 
-    static displayResidueDistribution(elementId: string, granularity: ResidueChartInterface["granularity"], rcsbId:string, facetLayoutGrid?:string[]): void {
+    static displayResidueDistribution(elementId: string, granularity: ResidueChartInterface["granularity"], rcsbId:string, facetLayoutGrid?:string[],chartDisplayConfig?: Partial<ChartDisplayConfigInterface>): void {
         createRoot(document.getElementById(elementId)).render(
-            <RcsbResidueChartComponent granularity={granularity} rcsbId={rcsbId} facetLayoutGrid={facetLayoutGrid}/>
+            <RcsbResidueChartComponent granularity={granularity} rcsbId={rcsbId} facetLayoutGrid={facetLayoutGrid} chartDisplayConfig={chartDisplayConfig}/>
         );
     }
 }